+++ src/spatVector.cpp
@@ -377,7 +377,7 @@ void SpatVector::reserve(size_t n) {
 void SpatVector::computeExtent() {
 	if (geoms.size() == 0) return;
 	extent = geoms[0].extent;
-	for (size_t i=1; i<geoms.size(); i++) {
+	for (size_t i=1; i<geoms.size(); i++) {	
 		extent.unite(geoms[i].extent);
 	}
 }
@@ -398,7 +398,6 @@ bool SpatVector::replaceGeom(SpatGeom p, unsigned i) {
 	if (i < geoms.size()) {
 		if ((geoms[i].extent.xmin == extent.xmin) || (geoms[i].extent.xmax == extent.xmax) ||
 			(geoms[i].extent.ymin == extent.ymin) || (geoms[i].extent.ymax == extent.ymax)) {
-
 			geoms[i] = p;
 			computeExtent();
 		} else {